requirements.txt
flask
requests
google-cloud-translate
python-dotenv
pytest
google-cloud-core